[STRUCTURAL] Extract Class: process diagnostics into internal/diag
First slice of Phase 3, which targets main.go as composition only. The
session log, structured event log and the stack dumps written on signal or
panic are process infrastructure, not composition: file layout,
permissions and swapping the standard logger's writer are none of main's
business.

Moves twelve declarations behind an exported API — StartSessionLogger,
StartEventLogger, BuildEventSink, WriteStackDump, WriteDiagnosticDump,
WritePanicDump, CaptureAllStacks, LogFileSuffix, the two logger types with
Path accessors, and the active-session-log accessors that a signal handler
on another goroutine reads.

The twenty tests covering them move with the code, adapted only for the
exported names, so the extraction is verified by the same assertions that
guarded it before rather than by new ones. main_unix_test.go is deleted —
its single test was the signal-dump case and now lives in
internal/diag/diag_unix_test.go behind the same build tag.

main.go: 3383 -> 3217 lines. No behavior change.
